Use `TargetInfo` from `mc-sgx-core-types`
Change to the  `TargetInfo` from `mc-sgx-core-types` in order to be able
to create a DCAP quote.
