+++ .github/workflows/ci.yml
@@ -24,7 +24,7 @@ permissions:
 jobs:
   build-dev:
     runs-on: [self-hosted, Linux, large]
-    container: mobilecoin/builder-install:v0.0.25
+    container: mobilecoin/builder-install:v0.0.27
 
     steps:
       - name: Check out code
@@ -42,7 +42,7 @@ jobs:
 
   build-prod:
     runs-on: [self-hosted, Linux, large]
-    container: mobilecoin/builder-install:v0.0.25
+    container: mobilecoin/builder-install:v0.0.27
 
     steps:
       - name: Check out code
@@ -63,7 +63,7 @@ jobs:
 
   build-and-test-wasm:
     runs-on: [self-hosted, Linux, large]
-    container: mobilecoin/builder-install:v0.0.25
+    container: mobilecoin/builder-install:v0.0.27
 
     steps:
       - name: Check out code
@@ -78,7 +78,7 @@ jobs:
 
   lint-rust:
     runs-on: [self-hosted, Linux, large]
-    container: mobilecoin/builder-install:v0.0.25
+    container: mobilecoin/builder-install:v0.0.27
 
     steps:
       - name: Check out code
@@ -95,7 +95,7 @@ jobs:
 
   build-and-test-go:
     runs-on: [self-hosted, Linux, large]
-    container: mobilecoin/builder-install:v0.0.25
+    container: mobilecoin/builder-install:v0.0.27
 
     steps:
       - name: Check out code
@@ -129,7 +129,7 @@ jobs:
 
   docs:
     runs-on: [self-hosted, Linux, large]
-    container: mobilecoin/builder-install:v0.0.25
+    container: mobilecoin/builder-install:v0.0.27
 
     steps:
       - name: Check out code
@@ -148,7 +148,7 @@ jobs:
 
   mc-tests:
     runs-on: [self-hosted, Linux, large]
-    container: mobilecoin/builder-install:v0.0.25
+    container: mobilecoin/builder-install:v0.0.27
 
     strategy:
       matrix:
@@ -191,7 +191,7 @@ jobs:
 
   consensus-tests:
     runs-on: [self-hosted, Linux, large]
-    container: mobilecoin/builder-install:v0.0.25
+    container: mobilecoin/builder-install:v0.0.27
 
     strategy:
       matrix:
@@ -224,7 +224,7 @@ jobs:
 
   fog-tests:
     runs-on: [self-hosted, Linux, large]
-    container: mobilecoin/builder-install:v0.0.25
+    container: mobilecoin/builder-install:v0.0.27
 
     strategy:
       matrix:
@@ -275,7 +275,7 @@ jobs:
 
   fog-ingest-tests:
     runs-on: [self-hosted, Linux, large]
-    container: mobilecoin/builder-install:v0.0.25
+    container: mobilecoin/builder-install:v0.0.27
 
     steps:
       - name: Check out code
@@ -298,7 +298,7 @@ jobs:
 
   fog-conformance-tests:
     runs-on: [self-hosted, Linux, large]
-    container: mobilecoin/builder-install:v0.0.25
+    container: mobilecoin/builder-install:v0.0.27
 
     steps:
       - name: Check out code
@@ -338,7 +338,7 @@ jobs:
 
   fog-local-network-tests:
     runs-on: [self-hosted, Linux, large]
-    container: mobilecoin/builder-install:v0.0.25
+    container: mobilecoin/builder-install:v0.0.27
 
     steps:
       - name: Check out code
@@ -525,7 +525,7 @@ jobs:
 
   minting-and-burning-tests:
     runs-on: [self-hosted, Linux, large]
-    container: mobilecoin/builder-install:v0.0.25
+    container: mobilecoin/builder-install:v0.0.27
 
     steps:
       - name: Check out code
