+++ .github/workflows/dependent-repos.yml
@@ -10,7 +10,7 @@ env:
 jobs:
   android-bindings:
     runs-on: [self-hosted, Linux, large]
-    container: mobilecoin/builder-install:v0.0.25
+    container: mobilecoin/builder-install:v0.0.26
 
     permissions:
       pull-requests: write
@@ -30,7 +30,7 @@ jobs:
 
   full-service:
     runs-on: [self-hosted, Linux, large]
-    container: mobilecoin/builder-install:v0.0.25
+    container: mobilecoin/builder-install:v0.0.26
 
     permissions:
       pull-requests: write
