+++ consensus/service/config/Cargo.toml
@@ -27,6 +27,6 @@ hex = "0.4"
 pem = "2.0"
 serde = { version = "1.0", default-features = false, features = ["alloc", "derive"] }
 serde_json = "1.0"
-serde_with = "2.3"
+serde_with = { version = "3.1", default-features = false }
 toml = "0.7"
