+++ light-client/verifier/Cargo.toml
@@ -11,7 +11,7 @@ base64 = "0.21"
 displaydoc = "0.2"
 prost = { version = "0.11", default-features = false, features = ["prost-derive"] }
 serde = { version = "1.0", default-features = false, features = ["alloc", "derive"] }
-serde_with = { version = "2.3", features = ["hex"] }
+serde_with = { version = "3.1", default-features = false, features = ["hex"] }
 
 mc-api = { path = "../../api" }
 mc-blockchain-types = { path = "../../blockchain/types" }
