+++ sgx/types/Cargo.toml
@@ -28,3 +28,4 @@ crate-type = ["rlib"]
 default = []
 
 [dependencies]
+mc-sgx-core-sys-types = "0.7.2"
